Factor Out API RPC Body
